FS-2: Renamed DTO folder to Entity
