+++ src/main/java/com/fleet/status/dao/IEventDAO.java
@@ -1,7 +1,6 @@
 package com.fleet.status.dao;
 
-import com.fleet.status.dto.Aircraft;
-import com.fleet.status.dto.Event;
+import com.fleet.status.entity.Event;
 
 import java.util.List;
