+++ src/main/java/com/fleet/status/dao/repository/AircraftRepository.java
@@ -1,6 +1,6 @@
 package com.fleet.status.dao.repository;
 
-import com.fleet.status.dto.Aircraft;
+import com.fleet.status.entity.Aircraft;
 import org.springframework.context.annotation.Profile;
 import org.springframework.data.repository.CrudRepository;
