+++ src/main/java/com/fleet/status/dao/repository/CarrierRepository.java
@@ -1,6 +1,6 @@
 package com.fleet.status.dao.repository;
 
-import com.fleet.status.dto.Carrier;
+import com.fleet.status.entity.Carrier;
 import org.springframework.context.annotation.Profile;
 import org.springframework.data.repository.CrudRepository;
