+++ src/main/java/com/fleet/status/dao/repository/EventRepository.java
@@ -1,6 +1,6 @@
 package com.fleet.status.dao.repository;
 
-import com.fleet.status.dto.Event;
+import com.fleet.status.entity.Event;
 import org.springframework.context.annotation.Profile;
 import org.springframework.data.repository.CrudRepository;
