+++ src/main/java/com/fleet/status/dao/repository/TypeRepository.java
@@ -1,6 +1,6 @@
 package com.fleet.status.dao.repository;
 
-import com.fleet.status.dto.Type;
+import com.fleet.status.entity.Type;
 import org.springframework.context.annotation.Profile;
 import org.springframework.data.repository.CrudRepository;
