+++ src/main/java/com/fleet/status/entity/Reason.java
@@ -1,4 +1,4 @@
-package com.fleet.status.dto;
+package com.fleet.status.entity;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import lombok.Data;
