+++ src/main/java/com/fleet/status/service/IAircraftService.java
@@ -1,6 +1,6 @@
 package com.fleet.status.service;
 
-import com.fleet.status.dto.Aircraft;
+import com.fleet.status.entity.Aircraft;
 
 import java.util.List;
