+++ src/main/java/com/fleet/status/service/impl/AircraftService.java
@@ -1,14 +1,12 @@
 package com.fleet.status.service.impl;
 
 import com.fleet.status.dao.impl.AircraftDAO;
-import com.fleet.status.dto.Aircraft;
+import com.fleet.status.entity.Aircraft;
 import com.fleet.status.service.IAircraftService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 
-import java.time.*;
-import java.time.format.DateTimeFormatter;
 import java.util.List;
 
 @Service
